backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
redis>=5.0.1
python-multipart>=0.0.9
httpx>=0.26.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
